perf(test): parallelise OCR fixtures and backend files for 40% faster runs
moonboard-ocr (135s → 35s): share a Tesseract scheduler across the 59 fixtures
per file via a new optional param on runOCR/parseWithProcessor, then run each
fixture with it.concurrent so we stop paying worker-spawn cost per test.

backend (188s → 62s): each vitest worker now owns its own Postgres database
(boardsesh_backend_test_w<POOL_ID>), so fileParallelism can be enabled without
TRUNCATEs stomping on each other. Schema creation moves to globalSetup (once
per run) and a new worker-db.ts rewrites DATABASE_URL at module-load time
before the app's db client caches its connection.

Full suite: 275s → 162s. No test logic changes; failing-before-me tests still
fail the same way.
