+++ packages/moonboard-ocr/src/__tests__/parser.canvas.test.ts
@@ -9,8 +9,9 @@
  * different image processing pipelines. Hold detection should be consistent.
  */
 
-import { describe, it } from 'vite-plus/test';
+import { describe, it, beforeAll, afterAll } from 'vite-plus/test';
 import path from 'path';
+import { createScheduler, createWorker } from 'tesseract.js';
 import { NodeCanvasImageProcessor } from './helpers/node-canvas-processor';
 import { parseWithProcessor } from '../parser';
 import { EXPECTED_RESULTS } from './fixtures/expected-results';
@@ -18,13 +19,28 @@ import { validateParseResult } from './helpers/test-utils';
 
 const FIXTURES_DIR = path.join(__dirname, 'fixtures');
 
+type Scheduler = Awaited<ReturnType<typeof createScheduler>>;
+
 describe('MoonBoard OCR Parser (Canvas Implementation)', () => {
+  let scheduler: Scheduler;
+
+  beforeAll(async () => {
+    scheduler = createScheduler();
+    const workerCount = 4;
+    const workers = await Promise.all(Array.from({ length: workerCount }, () => createWorker('eng')));
+    for (const w of workers) scheduler.addWorker(w);
+  }, 60_000);
+
+  afterAll(async () => {
+    await scheduler.terminate();
+  });
+
   for (const expected of EXPECTED_RESULTS) {
     describe(expected.fixture, () => {
-      it('should extract correct climb data', async () => {
+      it.concurrent('should extract correct climb data', async () => {
         const processor = new NodeCanvasImageProcessor();
         await processor.load(path.join(FIXTURES_DIR, expected.fixture));
-        const result = await parseWithProcessor(processor);
+        const result = await parseWithProcessor(processor, { scheduler });
 
         validateParseResult(result, expected, {
           validateOcr: true,
