+++ packages/moonboard-ocr/src/parser.ts
@@ -8,19 +8,20 @@
 import path from 'path';
 import { SharpImageProcessor } from './image-processor/sharp-processor';
 import type { MoonBoardClimb, ParseResult } from './types';
-import { parseWithProcessor, deduplicateClimbs } from './parser-core';
+import { parseWithProcessor, deduplicateClimbs, type ParseOptions } from './parser-core';
 
 // Re-export browser-safe core functions for backward compatibility
 export { parseWithProcessor, deduplicateClimbs };
+export type { ParseOptions };
 
 /**
  * Parse a single MoonBoard screenshot from file path (Node.js API).
  * This maintains backward compatibility with the original API.
  */
-export async function parseScreenshot(imagePath: string): Promise<ParseResult> {
+export async function parseScreenshot(imagePath: string, options: ParseOptions = {}): Promise<ParseResult> {
   const processor = new SharpImageProcessor();
   await processor.load(imagePath);
-  return parseWithProcessor(processor);
+  return parseWithProcessor(processor, options);
 }
 
 /**
